make feature extraction work on multiple gpu

+++ extract_features.py
@@ -1,10 +1,12 @@
 import os
 import tqdm
+import time
 import wandb
 import torch
 import hydra
 import shutil
 import datetime
+import subprocess
 import pandas as pd
 from PIL import Image
 from pathlib import Path
@@ -13,7 +15,7 @@ from torchvision import transforms
 
 from source.dataset import RegionFilepathsDataset
 from source.models import GlobalFeatureExtractor, LocalFeatureExtractor
-from source.utils import initialize_wandb, initialize_df, collate_region_filepaths
+from source.utils import initialize_wandb, initialize_df, collate_region_filepaths, is_main_process
 
 
 @hydra.main(
@@ -21,18 +23,36 @@ from source.utils import initialize_wandb, initialize_df, collate_region_filepat
 )
 def main(cfg: DictConfig):
 
-    run_id = datetime.datetime.now().strftime('%Y-%m-%d_%H_%M')
-    # set up wandb
-    if cfg.wandb.enable:
-        key = os.environ.get("WANDB_API_KEY")
-        wandb_run = initialize_wandb(cfg, key=key)
-        wandb_run.define_metric("processed", summary="max")
-        run_id = wandb_run.id
+    distributed = torch.cuda.device_count() > 1
+    if distributed:
+        torch.distributed.init_process_group(backend="nccl")
+        gpu_id = int(os.environ["LOCAL_RANK"])
+        if gpu_id == 0:
+            print(f"Distributed session successfully initialized")
+    else:
+        gpu_id = -1
+
+    if is_main_process():
+        print(f"torch.cuda.device_count(): {torch.cuda.device_count()}")
+        run_id = datetime.datetime.now().strftime("%Y-%m-%d_%H_%M")
+        # set up wandb
+        if cfg.wandb.enable:
+            key = os.environ.get("WANDB_API_KEY")
+            wandb_run = initialize_wandb(cfg, key=key)
+            wandb_run.define_metric("processed", summary="max")
+            run_id = wandb_run.id
+    else:
+        run_id = ""
+
+    if distributed:
+        obj = [run_id]
+        torch.distributed.broadcast_object_list(obj, 0, device=torch.device(f"cuda:{gpu_id}"))
+        run_id = obj[0]
 
     output_dir = Path(cfg.output_dir, cfg.experiment_name, run_id)
     slide_features_dir = Path(output_dir, "slide")
     region_features_dir = Path(output_dir, "region")
-    if not cfg.resume:
+    if not cfg.resume and is_main_process():
         if output_dir.exists():
             print(f"{output_dir} already exists! deleting it...")
             shutil.rmtree(output_dir)
@@ -54,53 +74,74 @@ def main(cfg: DictConfig):
             mini_patch_size=cfg.mini_patch_size,
             pretrain_vit_patch=cfg.pretrain_vit_patch,
             pretrain_vit_region=cfg.pretrain_vit_region,
+            verbose=(gpu_id in [-1, 0]),
         )
     elif cfg.level == "local":
         model = LocalFeatureExtractor(
             patch_size=cfg.patch_size,
             mini_patch_size=cfg.mini_patch_size,
             pretrain_vit_patch=cfg.pretrain_vit_patch,
+            verbose=(gpu_id in [-1, 0]),
         )
     else:
         raise ValueError(f"cfg.level ({cfg.level}) not supported")
 
     region_dir = Path(cfg.region_dir)
     slide_ids = sorted([s.name for s in region_dir.iterdir()])
-    print(f"{len(slide_ids)} slides with extracted patches found")
+    if is_main_process():
+        print(f"{len(slide_ids)} slides with extracted patches found")
 
     if cfg.slide_list:
         with open(Path(cfg.slide_list), "r") as f:
             slide_ids = sorted([Path(x.strip()).stem for x in f.readlines()])
-        print(f"restricting to {len(slide_ids)} slides from slide list .txt file")
+        if is_main_process():
+            print(f"restricting to {len(slide_ids)} slides from slide list .txt file")
 
-    process_list_fp = None
-    if (
-        Path(output_dir, f"process_list_{cfg.level}.csv").is_file()
-        and cfg.resume
-    ):
-        process_list_fp = Path(output_dir, f"process_list_{cfg.level}.csv")
+    df = initialize_df(slide_ids)
+    dataset = RegionFilepathsDataset(df, region_dir, cfg.format)
 
-    if process_list_fp is None:
-        df = initialize_df(slide_ids)
+    if distributed and is_main_process() and cfg.wandb.enable:
+        command_line = [
+            "python3",
+            "log_nproc.py",
+            "--output_dir",
+            f"{slide_features_dir}",
+            "--fmt",
+            "pt",
+            "--total",
+            f"{len(dataset)}",
+            "--log_to_wandb",
+            "--id",
+            f"{run_id}"
+        ]
+        subprocess.Popen(command_line)
+
+    time.sleep(5)
+
+    if distributed:
+        sampler = torch.utils.data.DistributedSampler(dataset, shuffle=True)
     else:
-        df = pd.read_csv(process_list_fp)
+        sampler = torch.utils.data.RandomSampler(dataset)
 
-    mask = df["process"] == 1
-    process_stack = df[mask]
-    total = len(process_stack)
-    already_processed = len(df) - total
-
-    region_dataset = RegionFilepathsDataset(df, region_dir, cfg.format)
-    region_subset = torch.utils.data.Subset(
-        region_dataset, indices=process_stack.index.tolist()
-    )
     loader = torch.utils.data.DataLoader(
-        region_subset, batch_size=1, shuffle=False, collate_fn=collate_region_filepaths
+        dataset,
+        sampler=sampler,
+        batch_size=1,
+        num_workers=cfg.num_workers,
+        shuffle=False,
+        drop_last=False,
+        collate_fn=collate_region_filepaths,
     )
 
-    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+    if gpu_id == -1:
+        device = torch.device(f"cuda")
+    else:
+        device = torch.device(f"cuda:{gpu_id}")
+    model = model.to(device, non_blocking=True)
 
-    print()
+    if is_main_process():
+        print()
+
     slide_ids, region_slide_ids = [], []
     slide_feature_paths, region_feature_paths = [], []
     x_coords, y_coords = [], []
@@ -109,11 +150,10 @@ def main(cfg: DictConfig):
         loader,
         desc="Slide Encoding",
         unit=" slide",
-        initial=already_processed,
-        total=total + already_processed,
         ncols=80,
         position=0,
         leave=True,
+        disable=not (gpu_id in [-1, 0]),
     ) as t1:
 
         with torch.no_grad():
@@ -131,6 +171,7 @@ def main(cfg: DictConfig):
                     ncols=80 + len(slide_id),
                     position=1,
                     leave=False,
+                    disable=not (gpu_id in [-1, 0]),
                 ) as t2:
 
                     for fp in t2:
@@ -145,7 +186,7 @@ def main(cfg: DictConfig):
                         img = img.to(device, non_blocking=True)
                         feature = model(img)
                         if cfg.save_region_features:
-                            save_path = Path(region_features_dir, f"{slide_id}_{x_y}.pt")
+                            save_path = Path(region_features_dir, f"{slide_id}_{x}_{y}.pt")
                             torch.save(feature, save_path)
                             region_feature_paths.append(save_path.resolve())
                             region_slide_ids.append(slide_id)
@@ -156,15 +197,8 @@ def main(cfg: DictConfig):
                 torch.save(stacked_features, save_path)
                 slide_feature_paths.append(save_path.resolve())
 
-                df.loc[idx, "process"] = 0
-                df.loc[idx, "status"] = "processed"
-                df.to_csv(
-                    Path(output_dir, f"process_list_{cfg.level}.csv"),
-                    index=False,
-                )
-
-                if cfg.wandb.enable:
-                    wandb.log({"processed": already_processed + i + 1})
+                if cfg.wandb.enable and not distributed:
+                    wandb.log({"processed": i + 1})
 
     slide_features_df = pd.DataFrame.from_dict({
         'feature_path': slide_feature_paths,
@@ -172,7 +206,13 @@ def main(cfg: DictConfig):
         'level': [f'{cfg.level}']*len(slide_ids),
         'tile_size': [cfg.region_size]*len(slide_ids),
     })
-    slide_features_df.to_csv(Path(output_dir, 'slide_features.csv'), index=False)
+
+    if distributed:
+        slide_features_csv_path = Path(output_dir, f'slide_features_{gpu_id}.csv')
+    else:
+        slide_features_csv_path = Path(output_dir, f'slide_features.csv')
+    slide_features_df.to_csv(slide_features_csv_path, index=False)
+
     if cfg.save_region_features:
         region_features_df = pd.DataFrame.from_dict({
             'feature_path': region_feature_paths,
@@ -182,9 +222,41 @@ def main(cfg: DictConfig):
             'x': x_coords,
             'y': y_coords,
         })
-        region_features_df.to_csv(Path(output_dir, 'region_features.csv'), index=False)
+        if distributed:
+            region_features_csv_path = Path(output_dir, f'region_features_{gpu_id}.csv')
+        else:
+            region_features_csv_path = Path(output_dir, f'region_features.csv')
+        region_features_df.to_csv(region_features_csv_path, index=False)
+
+    if distributed:
+        torch.distributed.barrier()
+        if is_main_process():
+            slide_dfs = []
+            if cfg.save_region_features:
+                region_dfs = []
+            for gpu_id in range(torch.cuda.device_count()):
+                slide_fp = Path(output_dir, f'slide_features_{gpu_id}.csv')
+                slide_df = pd.read_csv(slide_fp)
+                slide_dfs.append(slide_df)
+                os.remove(slide_fp)
+                if cfg.save_region_features:
+                    region_fp = Path(output_dir, f'region_features_{gpu_id}.csv')
+                    region_df = pd.read_csv(region_fp)
+                    region_dfs.append(region_df)
+                    os.remove(region_fp)
+            slide_features_df = pd.concat(slide_dfs, ignore_index=True)
+            slide_features_df = slide_features_df.drop_duplicates()
+            slide_features_df.to_csv(Path(output_dir, f'slide_features.csv'), index=False)
+            if cfg.save_region_features:
+                region_features_df = pd.concat(region_dfs, ignore_index=True)
+                region_features_df = region_features_df.drop_duplicates()
+                region_features_df.to_csv(Path(output_dir, f'region_features.csv'), index=False)
+
+    if cfg.wandb.enable and is_main_process() and distributed:
+        wandb.log({"processed": len(slide_features_df)})
 
 
 if __name__ == "__main__":
 
+    # python3 -m torch.distributed.run --standalone --nproc_per_node=gpu extract_features_dist.py --config-name 'debug'
     main()
